+++ src/capture/client/Capture/Output/TCPWriter/TCPWriter.h
@@ -28,6 +28,6 @@
 - (void)closeWriter;
 - (NSString *)getAbsolutePath:(NSString *)relativePath;
 - (void)createDirectory:(NSString *)relativePath;
-- (void)writeData:(NSData *)data relativePath:(NSString *)relativePath timestamp:(uint64_t)timestamp;
+- (void)writeData:(NSData *)data relativePath:(NSString *)relativePath timestamp:(Float64)timestamp;
 
 @end
